requirements.txt
environ==1.0
et_xmlfile==2.0.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
watchdog==6.0.0
xlrd==2.0.1
beautifulsoup4==4.12.3